refactor(AttendanceController): service 레이어 삭제
